+++ include/misc/ObjectNames.hpp
@@ -8,7 +8,7 @@ using namespace geode::prelude;
 
 class ObjectNames : public Singleton<ObjectNames> {
 public:
-    void checkNames();
+    void checkNames(bool ignoreVersion = false);
     Result<std::string_view> getName(unsigned int id);
     std::string deduceFromID(ZStringView id);
     const std::unordered_map<unsigned int, std::string>& getNames();